pyproject.toml
[project]
name = "ap-data-applications"
version = "0.1.0"
requires-python = "==3.12.8"
dependencies = [
    "google-genai",
    "pydantic_settings"
]